refactor(appeng): move getMonitor out of AEUtil
