+++ src/main/scala/li/cil/oc/integration/appeng/AEUtil.scala
@@ -1,10 +1,6 @@
 package li.cil.oc.integration.appeng
 
 import appeng.api.AEApi
-import appeng.api.storage.IMEMonitor
-import appeng.api.storage.data.IAEStack
-import appeng.me.GridAccessException
-import appeng.me.helpers.IGridProxyable
 import cpw.mods.fml.common.Loader
 import cpw.mods.fml.common.versioning.VersionRange
 import li.cil.oc.api
@@ -12,8 +8,6 @@ import li.cil.oc.common.item.Delegator
 import li.cil.oc.common.item.data.{DroneData, RobotData}
 import li.cil.oc.integration.Mods
 import net.minecraft.item.ItemStack
-
-import scala.reflect.ClassTag
 
 object AEUtil {
   val versionsWithNewItemDefinitionAPI = VersionRange.createFromVersionSpec("[rv2-beta-20,)")
@@ -164,17 +158,5 @@ object AEUtil {
       }
     }
     null
-  }
-
-  def getMonitor[T <: IAEStack[T] : ClassTag](controller: IGridProxyable): Option[IMEMonitor[T]] = {
-    for {
-      c <- Option(controller)
-      entry <- AEStackFactory.getEntry[T]()
-      inv <- try {
-        Option(c.getProxy.getStorage.getMEMonitor(entry.stackType).asInstanceOf[IMEMonitor[T]])
-      } catch {
-        case _: GridAccessException => None
-      }
-    } yield inv
   }
 }
